+++ activerecord/lib/active_record/connection_adapters/postgresql/oid/date_time.rb
@@ -7,7 +7,7 @@ module ActiveRecord
         class DateTime < Type::DateTime # :nodoc:
           def cast_value(value)
             case value
-            when "infinity" then ::Float::INFINITY
+            when "infinity", ::DateTime::Infinity then ::Float::INFINITY
             when "-infinity" then -::Float::INFINITY
             when / BC$/
               value = value.sub(/^\d+/) { |year| format("%04d", -year.to_i + 1) }